hpc/src/requirements.txt
torch>=2.3
torchvision>=0.18
pytorch-lightning>=2.3
numpy
opencv-python
pillow
pyyaml
tqdm
albumentations
pycocotools
shapely
geopandas
rasterio
rio-cogeo
faiss-gpu
open-clip-torch>=2.24.0
transformers>=4.41
accelerate
xformers
segment-anything
pymupdf
pytesseract
pdfminer.six
wandb
rich
scikit-image
pytorch-msssim
lpips
matplotlib
